user/requirements.txt
Flask==2.0.2
flask-pymongo==2.3.0
flask-jwt-extended==4.3.1
python-dotenv==0.19.2
werkzeug==2.0.2
pymongo==3.12.1
pymongo[srv]
cryptography
